+++ src/main/java/com/sprint/be_java_hisp_w23_g04/dto/response/BuyerDTO.java
@@ -1,9 +1,10 @@
-package com.sprint.be_java_hisp_w23_g04.dtoNew.response;
+package com.sprint.be_java_hisp_w23_g04.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 import lombok.AllArgsConstructor;
 import lombok.Data;
+import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 
 import java.util.List;
@@ -12,9 +13,10 @@ import java.util.List;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
+@EqualsAndHashCode(callSuper = false)
 @JsonInclude(JsonInclude.Include.NON_NULL)
-@JsonPropertyOrder({ "id", "name", "followed" })
-public class BuyerDTO extends UserDTO{
+@JsonPropertyOrder({"id", "name", "followed"})
+public class BuyerDTO extends UserDTO {
     private List<UserDTO> followed;
 
     public BuyerDTO(Integer id, String name, List<UserDTO> followed) {
