+++ src/main/java/com/sprint/be_java_hisp_w23_g04/dto/response/PostDTO.java
@@ -1,8 +1,8 @@
 package com.sprint.be_java_hisp_w23_g04.dto.response;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
 import lombok.Data;
+import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 
 import java.util.List;
@@ -10,8 +10,7 @@ import java.util.List;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class FilteredPostsDTO {
-    @JsonProperty("user_id")
-    private int userId;
+@EqualsAndHashCode(callSuper = false)
+public class PostDTO extends SellerDTO {
     private List<PostResponseDTO> posts;
 }
