--- src/main/java/com/sprint/be_java_hisp_w23_g04/dto/response/PostFollowersDTO.java
@@ -1,16 +1,0 @@
-package com.sprint.be_java_hisp_w23_g04.dto.response;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.*;
-
-import java.util.List;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class PostFollowersDTO {
-    @JsonProperty("user_id")
-    private Integer userId;
-    @JsonProperty("posts")
-    private List<PostResponseDTO> posts;
-}
