+++ src/main/java/com/sprint/be_java_hisp_w23_g04/dto/response/PostListDTO.java
@@ -2,19 +2,16 @@ package com.sprint.be_java_hisp_w23_g04.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
-import lombok.NoArgsConstructor;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 
 import java.util.List;
 
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class FollowedListDTO {
+public class PostListDTO {
     @JsonProperty("user_id")
     private Integer userId;
-    @JsonProperty("user_name")
-    private String userName;
-    @JsonProperty("followed")
-    private List<UserFollowDTO> followed;
+    private List<PostResponseDTO> posts;
 }
