+++ src/main/java/com/sprint/be_java_hisp_w23_g04/dto/response/ProductDTO.java
@@ -1,7 +1,9 @@
 package com.sprint.be_java_hisp_w23_g04.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
 @Data
 @NoArgsConstructor
