+++ src/main/java/com/sprint/be_java_hisp_w23_g04/dto/response/SellerDTO.java
@@ -1,17 +1,18 @@
-package com.sprint.be_java_hisp_w23_g04.dtoNew.response;
+package com.sprint.be_java_hisp_w23_g04.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
 import lombok.Data;
+import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 
 import java.util.List;
 
-
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
+@EqualsAndHashCode(callSuper = false)
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class SellerDTO extends BuyerDTO {
     private List<UserDTO> followers;
@@ -29,3 +30,4 @@ public class SellerDTO extends BuyerDTO {
         this.followersCount = followersCount;
     }
 }
+
