--- src/main/java/com/sprint/be_java_hisp_w23_g04/dto/response/UserFollowDTO.java
@@ -1,16 +1,0 @@
-package com.sprint.be_java_hisp_w23_g04.dto.response;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.AllArgsConstructor;
-import lombok.NoArgsConstructor;
-import lombok.Data;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class UserFollowDTO {
-    @JsonProperty("user_id")
-    private int userId;
-    @JsonProperty("user_name")
-    private String userName;
-}
