--- src/main/java/com/sprint/be_java_hisp_w23_g04/dtoNew/request/PostDTO.java
@@ -1,23 +1,0 @@
-package com.sprint.be_java_hisp_w23_g04.dtoNew.request;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.sprint.be_java_hisp_w23_g04.dtoNew.response.ProductDTO;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDate;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class PostDTO {
-    @JsonProperty("user_id")
-    protected int userId;
-    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "dd-MM-yyyy")
-    protected LocalDate date;
-    protected ProductDTO product;
-    protected int category;
-    protected double price;
-}
