--- src/main/java/com/sprint/be_java_hisp_w23_g04/dtoNew/response/SimpleMessageDTO.java
@@ -1,10 +1,0 @@
-package com.sprint.be_java_hisp_w23_g04.dtoNew.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class SimpleMessageDTO {
-    private String message;
-}
