+++ src/main/java/com/sprint/be_java_hisp_w23_g04/entity/Product.java
@@ -1,8 +1,8 @@
 package com.sprint.be_java_hisp_w23_g04.entity;
 
+import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import lombok.AllArgsConstructor;
 
 @Data
 @NoArgsConstructor
