--- src/main/java/com/sprint/be_java_hisp_w23_g04/entityNew/Post.java
@@ -1,24 +1,0 @@
-package com.sprint.be_java_hisp_w23_g04.entityNew;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDate;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class Post {
-    private Integer id;
-    @JsonProperty("user_id")
-    private Integer userId;
-    @JsonProperty("product_id")
-    private Integer productId;
-    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "dd-MM-yyyy")
-    private LocalDate date;
-    private Integer category;
-    private Double price;
-}
