--- src/main/java/com/sprint/be_java_hisp_w23_g04/entityNew/Product.java
@@ -1,17 +1,0 @@
-package com.sprint.be_java_hisp_w23_g04.entityNew;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class Product {
-    private Integer id;
-    private String name;
-    private String type;
-    private String brand;
-    private String color;
-    private String notes;
-}
