--- src/main/java/com/sprint/be_java_hisp_w23_g04/entityNew/User.java
@@ -1,24 +1,0 @@
-package com.sprint.be_java_hisp_w23_g04.entityNew;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.sprint.be_java_hisp_w23_g04.entity.Post;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.ArrayList;
-import java.util.List;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class User {
-    private Integer id;
-    private String name;
-    @JsonProperty("posts_id")
-    private List<Integer> postsId = new ArrayList<>();
-    @JsonProperty("followed_id")
-    private List<Integer> followedId = new ArrayList<>();
-    @JsonProperty("followers_id")
-    private List<Integer> followersId = new ArrayList<>();
-}
