+++ src/main/java/com/sprint/be_java_hisp_w23_g04/gateways/IPostGateway.java
@@ -1,7 +1,6 @@
 package com.sprint.be_java_hisp_w23_g04.gateways;
 
-import com.sprint.be_java_hisp_w23_g04.entityNew.Post;
-import com.sprint.be_java_hisp_w23_g04.entityNew.User;
+import com.sprint.be_java_hisp_w23_g04.entity.Post;
 
 import java.util.List;
