+++ src/main/java/com/sprint/be_java_hisp_w23_g04/gateways/IProductGateway.java
@@ -1,6 +1,6 @@
 package com.sprint.be_java_hisp_w23_g04.gateways;
 
-import com.sprint.be_java_hisp_w23_g04.entityNew.Product;
+import com.sprint.be_java_hisp_w23_g04.entity.Product;
 
 public interface IProductGateway {
