+++ src/main/java/com/sprint/be_java_hisp_w23_g04/gateways/IUserGateway.java
@@ -1,6 +1,6 @@
 package com.sprint.be_java_hisp_w23_g04.gateways;
 
-import com.sprint.be_java_hisp_w23_g04.entityNew.User;
+import com.sprint.be_java_hisp_w23_g04.entity.User;
 
 import java.util.List;
