+++ src/main/java/com/sprint/be_java_hisp_w23_g04/gateways/UserGatewayImp.java
@@ -1,6 +1,6 @@
 package com.sprint.be_java_hisp_w23_g04.gateways;
 
-import com.sprint.be_java_hisp_w23_g04.entityNew.User;
+import com.sprint.be_java_hisp_w23_g04.entity.User;
 import com.sprint.be_java_hisp_w23_g04.repository.IUserMediaRepository;
 import com.sprint.be_java_hisp_w23_g04.repository.UserMediaRepositoryImpl;
 import org.springframework.stereotype.Component;
