+++ src/main/java/com/sprint/be_java_hisp_w23_g04/repository/IProductMediaRepository.java
@@ -1,7 +1,7 @@
 package com.sprint.be_java_hisp_w23_g04.repository;
 
-import com.sprint.be_java_hisp_w23_g04.entityNew.Product;
-import com.sprint.be_java_hisp_w23_g04.entityNew.User;
+import com.sprint.be_java_hisp_w23_g04.entity.Product;
+import com.sprint.be_java_hisp_w23_g04.entity.User;
 
 import java.util.List;
