--- src/main/java/com/sprint/be_java_hisp_w23_g04/repository/ISocialMediaRepository.java
@@ -1,20 +1,0 @@
-package com.sprint.be_java_hisp_w23_g04.repository;
-
-import com.sprint.be_java_hisp_w23_g04.entity.User;
-
-import java.util.List;
-
-public interface ISocialMediaRepository {
-    List<User> findAllUsers();
-
-    User findUser(Integer userId);
-
-    void unfollowUser(int userId, int unfollowedUserId);
-
-
-    int getNextPostId(User user);
-
-    void savePost(User user);
-
-
-}
