+++ src/main/java/com/sprint/be_java_hisp_w23_g04/service/IUserMediaService.java
@@ -1,13 +1,13 @@
 package com.sprint.be_java_hisp_w23_g04.service;
 
-import com.sprint.be_java_hisp_w23_g04.dtoNew.response.SimpleMessageDTO;
-import com.sprint.be_java_hisp_w23_g04.dtoNew.response.BuyerDTO;
-import com.sprint.be_java_hisp_w23_g04.dtoNew.response.SellerDTO;
+import com.sprint.be_java_hisp_w23_g04.dto.response.SimpleMessageDTO;
+import com.sprint.be_java_hisp_w23_g04.dto.response.BuyerDTO;
+import com.sprint.be_java_hisp_w23_g04.dto.response.SellerDTO;
 
 import java.util.List;
 
 public interface IUserMediaService {
-    List<com.sprint.be_java_hisp_w23_g04.dtoNew.response.UserDTO> getAllUsers();
+    List<com.sprint.be_java_hisp_w23_g04.dto.response.UserDTO> getAllUsers();
 
     BuyerDTO getFollowedByUserId(Integer id, String order);
